requirements.txt
google-adk>=1.26.0
google-genai>=1.0.0
fastapi>=0.115.0
uvicorn>=0.34.0
websockets>=14.0
python-dotenv>=1.0.0
google-cloud-run>=0.10.0
